IConnectionInfoFeature.GetConnections returns most generic IEnumerable<ConnectionInfo> backed by list

+++ System.Net.Mqtt.Server/Features/IConnectionStatisticsFeature.cs
@@ -3,6 +3,6 @@ namespace System.Net.Mqtt.Server.Features;
 public interface IConnectionStatisticsFeature
 {
     long GetTotalConnections();
-    long GetActiveConnections();
+    int GetActiveConnections();
     long GetRejectedConnections();
 }
